Refactor AdminController into focused domain controllers
Split the monolithic AdminController into five focused controllers:
- AdminDashboardController: Handles dashboard analytics.
- AdminLeaguesController: Manages league-related operations.
- AdminRegistrationsController: Manages session registrations and payments.
- AdminSessionsController: Handles session-related operations.
- AdminUsersController: Manages user-related operations.

Introduced AdminControllerBase to centralize shared functionality
(e.g., IsAdminAsync) and reduce code duplication. Maintained existing
API routes to ensure compatibility with the Angular frontend.

Updated backend-refactor.json to mark the task as complete. Improved
logging and error handling for better debugging. No changes were made
to business logic or endpoint contracts, ensuring integration tests
remain functional.
